Remove ConstraintParameterParser
All checkers now use ConstraintStatementParameterParser, so we can
completely remove ConstraintParameterParser. A future commit will rename
ConstraintStatementParameterParser to ConstraintParameterParser, now
that the distinction between the two is obsolete.

This commit also updates several @uses test annotations that I forgot to
update in the respective commits that changed the tests. It also removes
an unused parameter that the tests for SingleValueChecker and
MultiValueChecker passed to their constructors (those constructors
declare no parameters).

The only remaining occurrences of ‘ConstraintParameterParser’ in the
repository are properties and methods which already use this term to
refer to ConstraintStatementParameterParser, in anticipation of the
aforementioned rename.

Change-Id: I4ac9717ac20f102d1dc2afc440cfc4b9af15ad56