Rename ConstraintStatementParameterParser to ConstraintParameterParser
Now that ConstraintParameterParser has been removed (see I4ac9717ac2),
ConstraintStatementParameterParser can use this shorter and more
suitable name. Most checkers and tests already used this name to refer
to the class in property and method names, so those members do not need
to be updated (the notable exception is ConstraintReportFactory).

Change-Id: I75118ead4f3c70f637be0bba63a9641c45fc2b4a