Split up range constraint violation messages
This splits up the violation messages for “range” and “diff within
range” constraints, both by range type (closed, left-open, right-open)
and, for “range”, by data type (“diff within range” ranges are always
quantity ranges, possibly with a time-type unit). For “diff within
range”, the old message is reused (without renaming it).

The half-open messages are phrased like “should be no less than” or
“should not be before” instead of “should be more than” or “should be
after” to make it clear that the range endpoints are inclusive.

Several tests are added to ensure that all possible messages are
covered. Some test methods are also removed to change the name from “too
small”/“too big” to “too early”/“too late” when it refers to a time.

Bug: T168268
Change-Id: I4f3fab270693736542c30ddd0faf4fde28844e26