Pass Constraint object into CheckResult
Replace the two constraintName and constraintId properties+parameters of
CheckResult with a single constraint property+parameter, which contains
this information and more. CheckResult::getConstraintName() delegates to
Constraint::getConstraintTypeName() for now; a future commit will
probably split it up into getConstraintTypeName() and …TypeQid() (which
may also be renamed to …TypeItemId()).

All places that instantiate a CheckResult are updated to pass the
correct parameters. Mostly, this is a simple search and replace of
$constraint->getConstraintTypeQid(), $constraint->getConstraintId() to
$constraint; the tests occasionally deviate from this pattern (and a few
instantiations are also broken across multiple lines).

This will allow us to add a constraint name parameter to Constraint (for
T167399) without also having to add it to CheckResult and therefore to
to all these instantiations.

Bug: T167399
Change-Id: Iec73ccf320ad4b6575cfee2524338a3a162e1823