Add link to constraint definition in gadget
Two new fields are added to the API description of a constraint, the
label of its type and the link to its definition. The gadget uses these
fields to render the title as a link instead of as plain text. The API
module gets three additional services injected so it can produce those
parameters.

The link to the definition is the link to the constraint statement
(group) for a constraint that was imported from a statement, and
otherwise the link to the property talk page, where the constraint
template is expected to be defined.

Bug: T167399
Change-Id: I9e2080950fc558e6083b5d69c33f86c4196e9caf