Improve link to constraint definition in special page
The special page gets a Config injected, so that it can get the property
ID of the ‘property constraint’ property, and uses that property ID
instead of the old, hard-coded constant. The link text is also changed
to look up the constraint item label if applicable (with fallback for
constraints imported from templates).

Bug: T167399
Change-Id: I511a9ceaab37127caa977402864962197317941a