Remove getConstraintName and getConstraintTypeName
The TODO comment in Constraint said that it should do a label lookup,
but since Constraint is supposed to be a value object, it’s much cleaner
to remove the method completely and leave the label lookup to the places
that use the constraint. The previous two changes already updated the
API (I9e2080950f) and special page (I511a9ceaab) accordingly; the only
remaining use of the methods is to order the constraints in the result,
which can just as well be done with the constraint type item ID, since
the main purpose of this sorting (according to I2dfabdd725) is to group
constraints of the same constraint type together, not to have the
constraint type groups in alphabetical order.

One test that compared the constraint type name in the CheckResult (in
FakeCheckerTest) is upgraded to compare the entire constraint instead,
since that is now recorded (as of Iec73ccf320); in another test, the
assertion for the constraint type name is simply removed, since the full
constraint object is already compared in the previous line.

Bug: T167399
Change-Id: Iec0f38a30b691a6a1f2001d8018b9ec64e961308