Add {class,relation}Parameter to ConstraintParameters
The ConstraintParameters test trait gains two methods to generate class
and relation parameters, and the tests for the “type” and “value type”
checkers are updated to use them. This means that the tests will
continue to work when we remove support for constraints imported from
templates.

Tests that were dedicated to constraint statements are removed, since
the new methods always use constraint statement-style parameters.
(ConstraintParameterParser tests ensure that both template and statement
parameters are supported.)

Bug: T172378
Change-Id: I507bbdeda6b638e7ae28cbc59a1ea9fb6cefa72d