Treat statement GUID as string, avoid parsing
Instead of parsing the statement GUID in the API and serializing it
again in DelegatingConstraintChecker, which is not a lossless
round-trip, pass it around as a string.

This is the correct solution for T165020, rather than the hacky
workaround in Icbe934f8b0.

DelegatingConstraintChecker gets a StatementGuidParser injected, since
it does need the parsed GUID to extract the entity ID part (to know
which entity to check). The API’s parseClaimIds is changed to use a
StatementGuidValidator, which is actually more strict than
StatementGuidParser, so some tests have to be updated to use a proper
GUID for the GUID part.

Bug: T165020
Change-Id: Ia9708d21871f8cf81896e63b3c39b571647e8379