Implement Format constraint with SPARQL
SparqlHelper gains a method to test if a string matches a regular
expression using SPARQL’s REGEX() function, and FormatChecker is
injected an optional SparqlHelper which it uses to check the constraint
if the SparqlHelper available.

The violation message gains parameters for the property, value and
regular expression.

The query service returns a different result for an invalid query, so we
can detect this case and report it as an invalid constraint parameter.
The SparqlHelper throws a ConstraintParameterException in that case, and
a new message is added for it.

ConstraintParameterRenderer’s formatByRole method is made public and
static, so that SparqlHelper can use it without needing a
ConstraintParameterRenderer injected.

The tests for the format checker are updated to no longer assert a TODO
status; the mock matchesRegularExpression implementation is the old
FormatChecker implementation which was removed in commit b325c00e99
(change I28d2850f89). A test for the behavior without SPARQL (TODO, as
before) is added.

Two tests for SparqlHelper are added, which test that it constructs the
query correctly (including correct input escaping) and that it correctly
detects the case of an invalid regex and throws an exception.

Bug: T102752
Change-Id: I920c6eedbfad9a75b39e01da595294c13c0b8cfc