Add ConstraintChecker::checkConstraintParameters method
This method returns a list of ConstraintParameterExceptions thrown by
ConstraintParameterParser. The implementation is fairly straightforward:
copy the invocations from the checkConstraint method and wrap them in
try-catch blocks.

The only complication is in RangeChecker, where the parameter parsing
depends on the type of the data (quantity or time). To solve this,
RangeChecker gets a PropertyDataTypeLookup injected, where it can look
up the data type of the property that the constraint belongs to.
(Strictly speaking, this is not exactly equivalent to the
checkConstraint method, which checks the data value type, but for
quantity and time types there is no difference between the two; the
documentation of parseRangeParameter() is amended to explicitly state
this.) ConstraintReportFactory also gets a PropertyDataTypeLookup
injected which it can pass to RangeChecker; the tests simply use a
lookup that wraps the already available EntityLookup.

The added tests for the new method each just check that the method
returns the right number of ConstraintParameterExceptions.

This new method constitutes the first part of T169531; later commits
will add a new API action to call this method, and make the gadget use
this API action to display parameter errors on the constraint statement.

Bug: T169531
Change-Id: I338577883d9bccfc2ea990f1ccc7173a3a503136