Add checkconstraintparameters API action
This new API action checks the validity of constraint parameters for a
given set of constraints, or for all constraints defined on a given set
of properties. The returned structure is meant to be extensible, as
later commits may add more information to ConstraintParameterException
objects (e. g. the affected parameter or the message key).

Note that this API only supports constraint IDs that are valid statement
GUIDs, i. e. constraints imported from statements.

Bug: T169531
Change-Id: I91f8b607ce1b37e5718f710058727caf5054166c