Add methods to check constraint parameters
This adds methods to check the constraint parameters of a particular
constraint, as well as of all constraints of a property, to
DelegatingConstraintChecker. These methods will form the backend of a
new API module to check constraint parameters.

Tests for the new methods are added, and the existing
DelegatingConstraintCheckerTest have to be tweaked slightly: the
EntityIdParser for the StatementGuidParser now needs to be able to parse
property IDs as well (in order to parse constraint statement GUIDs), and
all constraint IDs in the fake data have to be parsable constraint
statements as well.

Bug: T169531
Change-Id: Ic58ce9b4a00188955acb2f1bb1b17d5c9efbe642