Don’t check constraints on deprecated statements
A new CheckResult status is added, which indicates that a constraint
check was skipped because the subject statement is deprecated. All
checkers that check a generally “semantic” constraint (as opposed to a
more “technical” one) skip deprecated statements with this result status
right at the beginning of the checkConstraints() method. (Since
parameter parsing has not taken place – we’re not even interested in
incorrect parameters in this case – the parameters in the result are
always empty.)

All checker tests gain a test…ConstraintDeprecatedStatement test method.
For the updated checkers, it checks that the result status is
deprecated; for the other checkers, it checks that the constraint check
proceeds as expected, ignoring the statement rank. (The ResultAssertions
test trait gains a method to assert a deprecated check result.)

Bug: T170391
Change-Id: I8ec04faa05c0c4fa38137329de8104141a220d42
Depends-On: Id2372a0c640de517949e1af18951cfa2bbee8d67