Use NewItem+NewStatement in DiffWithinRangeCheckerTest
This makes the tests much more readable (the difference between the
tests is no longer hidden in the JSON files), and also lets us get rid
of the entity JSON files (it looks like some of them were already
unused). It is also easier to add tests of this kind in the future.

The values, statements and builders are static and shared between tests
as a slight optimization. (The builders are immutable, each modification
method creates a new clone.)

(Note that the new test code is not always exactly equivalent to the old
one – it often uses different values – but it still tests the same
thing. The exception is testDiffWithinRangeConstraintTooSmall, which, as
far as I can tell, actually used to test a too-big difference instead of
the intended too-small difference.)

Bug: T168240
Change-Id: I1c39e90d76a6d274d89e203888c07daea270f70c