Ignore deprecated statements in DiffWithinRangeChecker
When looking for the referenced statement of the constraint (e. g. the
“date of birth” statement for a constraint on “date of death”), skip
deprecated statements.

Two tests for the new behavior are added, along with a test for the
behavior when no referenced statement can be found, which was previously
untested (and should be identical to the behavior when only a deprecated
statement is present).

Bug: T170395
Change-Id: I614e32fc6200c45a4086777eab965f85045b1c57