Support properties with symmetric and inverse constraints
SymmetricChecker and InverseChecker used
ConnectionCheckerHelper::findStatement, which only supported item ID
values, not property ID values.

ConnectionCheckerHelper is essentially rewritten. It now has three
methods to find statements by different criteria, each of which returns
either the first matching statement or null to indicate that there is no
matching statement.

- findStatementWithProperty is the replacement for hasProperty, and is
  used by ItemChecker, TargetRequiredClaimChecker and
  ConflictsWithChecker if the items parameter is not set.
- findStatementWithPropertyAndEntityIdValue is the replacement for
  findStatement that is used by SymmetricChecker and InverseChecker. It
  supports property IDs as well as item IDs, but not somevalue and
  novalue, since this is not necessary for those checkers.
- findStatementWithPropertyAndItemIdSnakValues is the replacement for
  findStatement that is used by ItemChecker, TargetRequiredClaimChecker
  and ConflictsWithChecker if the items parameter is set.

Since the methods all return a statement and not an ItemIdSnakValue,
ItemIdSnakValue gains a method to turn a Snak into an ItemIdSnakValue.
This is used by ConflictsWithChecker to render the offending statement
into the violation message. All other checkers merely check if the
returned statement exists or not.

The test added in I6bc6711e00 is removed again, since none of the new
methods accept unparsed parameters.

Bug: T170626
Change-Id: Iac23aa144342148a028213afd16ffc175cabc1e7