Ignore deprecated statements for distinct values check
SparqlHelper::findEntitiesWithSameStatement gains a flag to optionally
ignore deprecated statements with the same value. UniqueValueChecker
sets that flag, since the deprecated statement is already marked as
incorrect, so there’s no need to report a constraint violation.

To filter out statements by rank, we need the wikibase: prefix, which is
a bit more complicated than other prefixes: the real prefix according to
RdfVocabulary is currently <http://wikiba.se/ontology-beta#>, and this
is also the RDF exports and dumps. However, in the Wikidata Query
Service, the prefix is replaced with <http://wikiba.se/ontology#>
(without the -beta) by the live updater and by the munger that
pre-processes data dumps before they are loaded. In order to support
both WDQS and local setups without live updater and without munger, we
filter out deprecated rank statements using both prefixes.

Bug: T170403
Change-Id: Icdaa469d9a5f2bdf7f0ba1a8dd3c2c331791b16c