Extract ConstraintReportPanel element into ui module
The code to create an OO.ui PanelLayout for a constraint report is
extracted from gadget.js into a new OO.ui class, ConstraintReportPanel,
which resides in the ui module (since it can be used independently of
the gadget).

The idea is to similarly extract the StackLayout into a new element
(probably ConstraintReportStack), and to implement T169971 there. The
gadget is then left only with the task of creating these elements and
appending them to statements.

Change-Id: Ic7143092eb91cc47e98b50cb4538745ef0e01c76