Use group label as popup label
The buildPopup() $content argument is now always the $element of a
ConstraintReportList, and all ConstraintReportGroups are initialized
with a label. Instead of adding a separate label to the popup content,
the buildPopup() function moves the label of the first constraint report
group into the head area of the popup content. (A label inside that head
area gets some special padding, which is why we don’t just leave the
label where it is.)

Change-Id: Ieedfb8bdf51c683cf05c53a96d8ce73559bd7584