Add visual distinction between groups in a list
This adds a border and some vertical space between groups in a
constraint report list. The border is slightly more pronounced than the
border between individual constraint report panels.

The border is added as border-bottom on all but the last group, instead
of as border-top on all but the first (i. e., with the vertical space
selector), because otherwise the border would appear below the vertical
space (the margin is not part of the element). The alternative solution
to this would be to use padding instead of margin for the vertical
space, but a margin makes more sense semantically.

Change-Id: If025d23f072969f69c93188363117dc4b4b4c705