Distinguish between non-mandatory and mandatory constraints
This adds a new constraint status, “warning”, for all violations of
constraints that are not mandatory. “violation” is only used for
violations of constraints that are explicitly marked mandatory, using
the “constraint status: mandatory constraint” constraint parameter
(which ConstraintParameterParser is taught to parse).

To avoid having to change all checkers and duplicating the constraint
status parsing into them, DelegatingConstraintChecker downgrades
“violation” to “warning” if appropriate, so that the individual checkers
remain unchanged. A few tests that include DelegatingConstraintChecker
are adjusted to make their constraints mandatory so that the result is
still the expected “violation”.

On the special page, warnings are displayed in yellow; the status
“exception”, which previously used this color, is changed to a slightly
darker green than “compliance”. The gadget is slightly updated to
include warnings in the report, but the full update for the gadget
(different texts, different icon, …) will happen in a separate commit.

Bug: T164254
Change-Id: I6fa0af86dd7d3d00abc4cd004417667028599428