Log details of long-running constraint checks
A new helper class is added, which takes care of logging decisions.
Currently, it logs constraint checks that take too long, because we
might be interested in the details of those constraint checks to
investigate why they took so long. The message is logged as “info” or
“warning”, depending on the duration; both limits (whether to log at
all, and whether to bump the level to “warning”) are configurable, and
can be set to null to disable the logging completely.

Bug: T171281
Change-Id: Icec9925453c1bc75a5be9ee033b86e44cf6790a9