Report unimportable constraint parameters
When the constraint parameters could not be imported, an "@error"
parameter is stored instead (see I52e842e50b for details). This commit
teaches ConstraintParameterParser and DelegatingConstraintChecker to
check for this parameter, so that it is indicated on the special page,
in the wbcheckconstraints API, and in the wbcheckconstraintparameters
API. Two messages are added for this case: one is used when the
ConstraintRepository detected that the parameter were too long, and the
other is a generic message for any other (unknown) error.

Bug: T171479
Change-Id: Ia11092746c895544fa68127289254185656650f2