Don’t store overly long constraint parameters
The constraint_parameters column of the wbqc_constraints table has type
TEXT, which is limited to 65535 bytes. If we attempt to store any longer
constraint parameters in it (for example, from a constraint with a very
long exception list), the database will either throw an error or
silently truncate the constraint parameters, depending on configuration.
We don’t want either of that to happen, so check the length (in bytes)
of the encoded constraint parameters before they are stored and replace
them with an error object if they’re too long.

If the constraint parameters still can’t be decoded for some reason,
also record an error object, but without the truncation indicator (since
we have no idea why the constraint parameters are invalid).

A second commit (Ia11092746c) teaches the rest of the extension to
recognize the error objects and properly report the error when a
constraint check happens.

Bug: T171479
Change-Id: I52e842e50b4fd60fbaacf9455a46bc031b982d0c