Fix classes without namespace
WikibaseQualityConstraintsHooks is moved to the
WikibaseQuality\ConstraintReport namespace; as far as I can tell,
there’s no reason it shouldn’t be there. (Note that you’ll need to run
`composer install` to update the classmap.)

tests/bootstrap.php can be removed completely. It is an exact copy of
the MediaWiki core bootstrap.php file (except with $IP instead of
__DIR__, and some whitespace changes), and no other extension outside of
WikibaseQuality seems to have this file. (Some have bootstrap files for
their tests, but with different content, but I don’t think we need to
redefine Core’s bootstrapping.)

This lets us drop an exception from the Wikibase phpcs ruleset.

Change-Id: I6301ec265b9816ded9ce89456ca8806cb978149c