Add Context interface and StatementContext implementation
A Context is a container for all the information that a
ConstraintChecker needs to check a constraint, apart from the constraint
itself. A context also knows where to store constraint check results in
an API response.

The StatementContext implementation matches the current API output
format. A later commit will add another implementation with a more
extensible API output format, which is required for T168532. It will
then be possible to switch between the API output formats by changing
the Context implementation, which should be possible to do via a
configuration change, or perhaps even via an API parameter.

The StatementContext implementation derives from an AbstractContext base
class, which contains some common functionality that other context
implementations can also use. Possible Context implementations include:

- contexts for qualifier and reference snaks, for T168532
- contexts for snaks that have not yet been saved, for T168626
- contexts for whole entities and no snak, for T163683

Bug: T171743
Change-Id: Ifec159629b6b07d0a1fd6e79c593c34989429142