Move constraint check tracking to LoggingHelper
LoggingHelper gets a statsd data factory inserted and is now responsible
for statsd tracking as well as structured logging. The tracking of
constraint check duration is moved from  the DelegatingConstraintChecker
to the LoggingHelper.

Since LoggingHelper already has the short class name of the constraint
checker, that is also included in the tracked metric, which lets us show
just that short name in Grafana (instead of the constraint type item ID,
which is less useful.)

Bug: T171285
Change-Id: I2be54579238e7b92b0a8a384b0214c8142c029f1