Support property IDs in UniqueValueChecker
UniqueValueChecker is not restricted to items – it can also find a
statement with the same value on properties, so it should also be able
to render property IDs into the violation message. formatItemIdList
doesn’t support that, so add a new function formatEntityIdList to
ConstraintParameterRenderer. (Since SparqlHelper can return null list
elements if a query result couldn’t be parsed into an entity ID, the
method also skips null elements.)

Bug: T172007
Change-Id: Ia174dfb315e10163862d5e78273eef61a4e16a3d