Fix type checking SPARQL fallback
The SPARQL fallback needs to happen at the outermost level, with the
original entity, not on some inner nested entity. Consider the following
subclass hierarchy:

                v---------<
    Q1 -> Q2 -> Q3 -> Q4 -^
           \
            > Q5

The PHP subtype checker (isSubclassOf()) will run through the Q3<->Q4
cycle until it runs into the check limit. (This could of course be
optimized, but that’s beside the point: there are enough hierarchies on
Wikidata that exceed the limit without cycles.) Previously, it would
then fall back to SPARQL, asking it if Q3 or Q4 are subclasses of the
desired class (Q5). This is incorrect, since the entity we’re actually
interested in is Q1.

The solution to this is to perform the fallback at an outer level,
outside of the recursive type check. To this end, a new method
isSubclassOfWithSparqlFallback() is added, which performs this fallback,
while isSubclassOf() throws an exception to indicate the fallback case
instead of returning false. hasClassInRelation() and all the tests then
use this new function instead of calling isSubclassOf() directly.

Bug: T169326
Change-Id: I2aaf3d6fa69791eff71718e808085cc99c128595