Use Context interface everywhere
Functions that previously accepted a statement, a constraint and an
entity (chiefly, ConstraintChecker::checkConstraint) are changed to
accept a context and a constraint instead. All the information related
to the entity or statement is obtained from the context; checkers that
actually need a statement (and not just snak data) also check that the
context is a statement-type context, and otherwise silently abort.

A CheckResult contains a context instead of an entity and a statement.
Some CheckResult accessor functions are updated to get the information
from the context (and occasionally renamed to a more appropriate name);
others are removed completely: specifically, we don’t want to have
getPropertyId() on CheckResult because it’s ambiguous between the
property ID of the statement containing the context and the property ID
of the snak being checked.

The API uses Context::storeCheckResultInArray() to store the result in
the response array. This means that changing the API output format is as
simple as switching the Context implementation.

Since the vast majority of this change is completely mechanical
replacements, I wrote a script that produces the diff, which may be
found at [1]. (The script also produces the non-trivial parts of the
diff, because I found it very comforting to have *everything* in that
script, so that I could easily `git reset --hard` and restore the
change. Perhaps that also makes it easier to rebase this change.)

[1]: https://gist.github.com/lucaswerkmeister/59e40bb3053c2f6eaaf53c69a9fdea79

Bug: T171743
Change-Id: I04c34ea7c0cd04d213b34faa9016dd3ff813b88c