Skip novalue/somevalue instead of reporting violation
Checkers that use the value of a statement previously reported a
violation for novalue/somevalue statements. This is rarely appropriate,
since semantically, those values should usually be accepted – the
constraints simply don’t apply in those cases. The checkers are updated
to return a compliance result in those cases instead of a violation, and
tests for the old behavior are updated to expect compliance instead.

DiffWithinRangeChecker also reported a violation if the statement of the
referenced property was novalue/somevalue; instead, it skips those
statements now, and either uses the first value statement it finds or
reports a violation due to missing statement otherwise. Tests are added
for this behavior.

Tests for CommonsLinkChecker and FormatChecker used to use novalue
statements to test that those checkers still check deprecated
statements; they have been changed to check actual statements that
violate constraints.

Two messages that are no longer used (generic “must have a value” and
DiffWithinRangeChecker-specific “referenced property must have a value”)
are removed.

Bug: T172130
Change-Id: I8bb33ead10770c938d04b00dd637924d4599a752