Improve rangeParameter helper and use it in RangeCheckerTest
All the remaining RangeCheckerTest methods that used template-style
constraint parameters are changed to use the rangeParameter helper of
the ConstraintParameter test trait. To simplify the test code, that
helper is improved to support specifying time values without having to
instantiate a TimeValue. (The code to do this is partly taken from
ConstraintParameterParser.)

Bug: T172378
Change-Id: Ie3a32b648f3a5d0fa2d2a741323242258a9398a5