Migrate ConnectionChecker tests to statement parameters
A new method for an items parameter is added to the ConstraintParameters
test trait, and the tests for ConnectionChecker use propertyParameter
and itemsParameter instead of specifying template-style constraint
parameters directly.

Tests that were dedicated to constraint statements are removed, since
they are now redundant – the trait methods always use constraint
statements, and ConstraintParameterParser tests ensure that, for now,
both template parameters and statement parameters are supported.

Bug: T172378
Change-Id: I4c51e9c43829b386770eafb4cff18ccaf34b1678