Remove constraint_status from API tests
Instead of setting the constraint_status of a constraint so that a
violation is not downgraded to a warning, expect warnings. This makes
the API tests independent of statement vs. template parameters.

See change I6fa0af86dd / commit 79c9108fe3 / task T164254 for details on
the downgrading that we now no longer avoid.

Bug: T172378
Change-Id: Id80a7ea59b7d626894fa9a2f3325aa6b485fc24c