Migrate DelegatingConstraintCheckerTest to statement params
Two new helper functions, for constraint exceptions and constraint
status, are added to the ConstraintParameter test trait, and
DelegatingConstraintCheckerTest gains a tiny helper function to resolve
a constraint type name to item ID.

The constraint data that the test loads into the database is updated to
always use statement-style constraint parameters and constraint type
item IDs instead of strings. This ensures that these tests keep working
after support for constraint templates is removed.

Bug: T172378
Change-Id: I548c02ab550ced0f3588cf00b411727fb16c6ef9