Add MainSnakContext implementation
This holds the same information as StatementSnak, but uses a different
result layout that can also accommodate qualifier and reference
constraint check results, as well as other information on a property.
Most of the implementation of the layout is in an intermediate abstract
class which can also be used by qualifier and reference contexts.

Change-Id: I4e0808023787a2e606a669d20d62e6df77fad877