Add support for parsing syntax clarification param
For the Format constraint, “syntax clarification” qualifiers (constraint
parameters) can explain the regular expression in multiple languages.
This commit adds methods to ConstraintParameterParser to parse these
qualifiers and extract the syntax clarification for a given language,
including fallback language support.

Note that the parseMultilingualTextParameter() helper method does
essentially the same as the MultilingualTextValue constructor, but
without using that class. This is so we can provide better error
messages for the two error cases. However, if the multilingual text
datatype ever becomes available on Wikidata, the syntax clarification
parameter should be converted to that type.

Bug: T170374
Change-Id: Iac5675f57ae173370b366785c22aafa533a2d2b3