Include syntax clarification in Format violation message
If a syntax clarification in the user’s language is available, the
Format checker emits a different message which mainly features the
clarification text and only includes the regex in parentheses. This
hopefully makes the violation easier to understand and fix for users who
aren’t familiar with regular expressions. (Also, in the existing
message, “pattern” is changed to “regex”, so that users at least have
something to search for when no clarification is available.)

In the FormatChecker tests, the existing violation assertions are also
updated to assert that the message matches the message without syntax
clarification, since no syntax clarification is provided in those tests.
A new test is added with syntax clarification to test the new message,
including a helper method for a syntax clarification parameter.

Bug: T170374
Change-Id: Ie20eebbdf7a18a9a2642599d8daa904dadc1adc0