Make SparqlHelper timeout detection configurable
SparqlHelper’s timeout detection (used to count timeouts in statsd) is
extracted into a separate function, and a configuration variable is
added for the exception class(es) that it looks for in the error
message.

In addition to the old timeout exception class, a second value is added
to the default value of that configuration variable, which reflects the
recently changed behavior of the Wikidata Query Service.

Tests for the new function are also added.

Bug: T173368
Change-Id: I8d4fde6fa03b9036380aaf505713cdaf7433d571