Fix constraint_status result parameter
For mandatory constraints, there should be a constraint_status parameter
in the constraint result array. I think the previous code was supposed
to be `$result->getParameters()['…'] = …`, but that wouldn’t have worked
either, since getParameters() returns a copy of the parameters array,
not a reference to it.

The test for a mandatory constraint is updated to also check for the
presence of the constraint_status parameter, and also gets some general
improvements.

Change-Id: I5f3026ac0bcb0bac7dde9fe64812f73e28f3ada6