Migrate SpecialConstraintReportTest to statement parameters
Use constraint type item IDs from the default configuration instead of
constraint type strings. This requires that we instantiate the
ConstraintReportFactory ourselves so we can inject our config into it,
which is unfortunately a rather unwieldy instantiation (mostly copied
from ConstraintReportFactory::getDefaultInstance(), except that we use a
PlainEntityIdFormatter).

Instead of making the constraint mandatory, we expect a warning
(DelegatingConstraintChecker downgrades non-mandatory violations to
warnings).

Bug: T172378
Change-Id: I26c0abce9a441b22ac8edf0d783075ff20233f60