Fix and document test setup
The bootstrap file was removed in I6301ec265b (I wasn’t aware that it
was referenced in the phpunit file). It seems that it can simply be
removed and the tests still work (since we’re already running
MediaWiki’s phpunit.php file, not phpunit directly).

The instructions to build with the configuration file are based on the
Travis build script (see build/travis/script.sh).

Bug: T165616
Change-Id: I2994ef4986c35a937655e90191ce06a6cf89eb56