Remove src_dir from .coveralls.yml
coveralls (specifically: vendor/bin/coveralls -v) is run as part of the
Travis build (see build/travis/after_script.sh) and complains that
src_dir is an unrecognized option. Presumably it was removed at some
point in the past; it seems that coveralls is still able to find the
source files without the option.

Bug: T165617
Bug: T165616
Change-Id: I8d8367037238e7b0f9e75d3849fcdc2c4cf5739c