Remove support for linking to property talk pages
Since there are no more constraints imported from templates, all
constraint IDs should be valid statement GUIDs, and we can remove the
fallback of linking to the property talk page.

This requires some minor changes to the API test, which now needs to
supply a parseable statement GUID as constraint ID, along with a
statement GUID parser that can parse property IDs as the first part.

Bug: T171291
Change-Id: I7524c1e2166f54eb15f430dfb2ee853173e40ccf