Add QualifierContext implementation and tests
QualifierContext is a context for a qualifier of a statement, deriving
from the abstract APIv2 context (just like MainSnakContext). It
shouldn’t be used yet, because I think some constraint checkers should
be updated to use more information from qualifier contexts (for
instance, I suspect the correct interpretation for “conflicts with” in a
qualifier is to look at the other snaks of the qualifier, not the other
statements of the entity), but that’s not a problem with the context
implementation itself, so I think that can be merged without problems.

Bug: T168532
Change-Id: Ibeaa9a2b0dd2c0a6269aae2922227d6ab0bf9c78