Reduce phpcs line length limit to 140
Some recent changes seem to have removed most of the overlong lines – I
think that’s partly due to the introduction of Context (which
significantly shortens all instantiations of CheckResult) and partly due
to the removal of constraint template support. The remaining lines
longer than 140 characters can be wrapped (or, in one case, a value
extracted) fairly simply, which is done in this commit.

(In new code, I try to stay below 100 characters nowadays, but for
existing code we’re still far away from that.)

Change-Id: Ideb26856c2769463dc7ccea704fa14b95e9da6e8