Cache regex check results
SparqlHelper::matchesRegularExpression is changed to be a wrapper around
a new function containing the previous code, caching its result. This
should hopefully reduce the number of requests to the query service.

Currently, there are 1786 format constraints defined on Wikidata, for a
total of 77,862,938 format constraint checks on statements. This was
determined with a script that may be found in paste P5921, and does not
include the statement count for the property P2093 (author name string),
which cannot be counted on WDQS. According to SQID, there are 24,150,684
statements for that property [1], so the total count is approximately
one hundred million format check results which might be cached.
(However, most of these checks will probably never be performed since
the corresponding items are rarely visited, and the set of checks that
are performed frequently is likely much smaller.)

To avoid putting arbitrary user input into the cache key, both regex and
text are hashed. If desired, the hash for the regex can be “reversed”
with a WDQS query like the following:

    SELECT ?property ?regex WHERE {
      ?property p:P2302 [
        ps:P2302 wd:Q21502404;
        pq:P1793 ?regex
      ].
      FILTER(SHA256(?regex) = "...")
    }

When the property is known, the hash for the value can similarly be
“reversed”, as long as the property doesn’t have too many statements and
isn’t a “Commons link” property (those map to URIs on the query
service).

[1]: https://tools.wmflabs.org/sqid/#/view?id=P2093

Bug: T173696
Change-Id: Iaaac950b483aaff83aa13b9f1b7d5090cd6c627f